:hammer: Refactoring routing-related code
